refactor: organize DTOs into user/goal/milestone subdirectories for scalability
