+++ backend/src/main/java/com/calendar/milestone/controller/dto/request/milestone/MilestoneRequest.java
@@ -1,4 +1,4 @@
-package com.calendar.milestone.controller.dto.request;
+package com.calendar.milestone.controller.dto.request.milestone;
 
 
 import java.time.LocalDate;
