+++ backend/src/main/java/com/calendar/milestone/controller/dto/request/user/UserPostRequest.java
@@ -1,18 +1,16 @@
-package com.calendar.milestone.controller.dto.response;
+package com.calendar.milestone.controller.dto.request.user;
 
 import java.time.LocalDate;
-import java.time.LocalDateTime;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.Pattern;
 import jakarta.validation.constraints.Size;
 
-public class UserResponse {
 
-    @JsonProperty("id")
-    private Integer id;
+public class UserPostRequest {
 
     @JsonProperty("name")
     @NotNull
@@ -32,18 +30,13 @@ public class UserResponse {
     @Size(max = 30)
     private String email;
 
-    @JsonProperty("created_at")
-    private LocalDateTime createdAt;
-
-    @JsonProperty("updated_at")
-    private LocalDateTime updatedAt;
-
-    @JsonProperty("deleted_at")
-    private LocalDateTime deletedAt;
-
-    public Integer getId() {
-        return id;
-    }
+    @JsonProperty("password")
+    @NotNull
+    @Size(min = 8, max = 20, message = "Password must be between 8 and 20 characters.")
+    @Pattern(
+            regexp = "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[!@#$%^&*()_+\\[\\]{};':\"\\\\|,.<>/?]).{8,20}$",
+            message = "Password must contain uppercase, lowercase, digit, and special character.")
+    private String password;
 
     public String getName() {
         return name;
@@ -61,20 +54,8 @@ public class UserResponse {
         return email;
     }
 
-    public LocalDateTime getCreatedAt() {
-        return createdAt;
-    }
-
-    public LocalDateTime getUpdatedAt() {
-        return updatedAt;
-    }
-
-    public LocalDateTime getDeletedAt() {
-        return deletedAt;
-    }
-
-    public void setId(int id) {
-        this.id = id;
+    public String getPassword() {
+        return password;
     }
 
     public void setName(String name) {
@@ -93,15 +74,7 @@ public class UserResponse {
         this.email = email;
     }
 
-    public void setCreatedAt(LocalDateTime createdAt) {
-        this.createdAt = createdAt;
-    }
-
-    public void setUpdatedAt(LocalDateTime updatedAt) {
-        this.updatedAt = updatedAt;
-    }
-
-    public void setDeletedAt(LocalDateTime deletedAt) {
-        this.deletedAt = deletedAt;
+    public void setPassword(String password) {
+        this.password = password;
     }
 }
