+++ backend/src/main/java/com/calendar/milestone/controller/dto/response/user/UserResponse.java
@@ -1,16 +1,18 @@
-package com.calendar.milestone.controller.dto.request;
+package com.calendar.milestone.controller.dto.response.user;
 
 import java.time.LocalDate;
+import java.time.LocalDateTime;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Pattern;
 import jakarta.validation.constraints.Size;
 
+public class UserResponse {
 
-public class UserPostRequest {
+    @JsonProperty("id")
+    private Integer id;
 
     @JsonProperty("name")
     @NotNull
@@ -30,13 +32,18 @@ public class UserPostRequest {
     @Size(max = 30)
     private String email;
 
-    @JsonProperty("password")
-    @NotNull
-    @Size(min = 8, max = 20, message = "Password must be between 8 and 20 characters.")
-    @Pattern(
-            regexp = "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[!@#$%^&*()_+\\[\\]{};':\"\\\\|,.<>/?]).{8,20}$",
-            message = "Password must contain uppercase, lowercase, digit, and special character.")
-    private String password;
+    @JsonProperty("created_at")
+    private LocalDateTime createdAt;
+
+    @JsonProperty("updated_at")
+    private LocalDateTime updatedAt;
+
+    @JsonProperty("deleted_at")
+    private LocalDateTime deletedAt;
+
+    public Integer getId() {
+        return id;
+    }
 
     public String getName() {
         return name;
@@ -54,8 +61,20 @@ public class UserPostRequest {
         return email;
     }
 
-    public String getPassword() {
-        return password;
+    public LocalDateTime getCreatedAt() {
+        return createdAt;
+    }
+
+    public LocalDateTime getUpdatedAt() {
+        return updatedAt;
+    }
+
+    public LocalDateTime getDeletedAt() {
+        return deletedAt;
+    }
+
+    public void setId(int id) {
+        this.id = id;
     }
 
     public void setName(String name) {
@@ -74,7 +93,15 @@ public class UserPostRequest {
         this.email = email;
     }
 
-    public void setPassword(String password) {
-        this.password = password;
+    public void setCreatedAt(LocalDateTime createdAt) {
+        this.createdAt = createdAt;
+    }
+
+    public void setUpdatedAt(LocalDateTime updatedAt) {
+        this.updatedAt = updatedAt;
+    }
+
+    public void setDeletedAt(LocalDateTime deletedAt) {
+        this.deletedAt = deletedAt;
     }
 }
